% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboAPI.R
\name{all_exports}
\alias{all_exports}
\title{Downloads forms from KPI}
\usage{
all_exports(user, pwd = "", api = "https://kobo.humanitarianresponse.info")
}
\arguments{
\item{user}{Optional. A single string indicating the username}

\item{pwd}{Password of the Kobo account to use.}

\item{api}{The URL at which the API can be accessed. Default to "kobo.humanitarianresponse.info"}
}
\value{
Returns a dataframe with all the exports available.
}
\description{
Returns a dataframe with all the exports available.
}
\author{
Elliott Messeiller
}
