% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboAPI.R
\name{api_type}
\alias{api_type}
\title{Check the API type}
\usage{
api_type(api = "https://kobo.humanitarianresponse.info")
}
\arguments{
\item{api}{The URL at which the API can be accessed.}
}
\value{
API type
}
\description{
Check the API type
}
\author{
Elliott Messeiller
}
