% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate_api}
\alias{authenticate_api}
\title{Authentication for kobo API}
\usage{
authenticate_api(auth_type, auth_key = NULL, api)
}
\arguments{
\item{auth_type}{(character) Two types of authentication are possible:
- `client`: use the `auth_key` `user:password` to authenticate to the kobo API
- `token`: uses the `token` provided in `auth_key` to authenticate}

\item{auth_key}{(character) Can either be:
- A combination of the kobo and password in format: `{user}:{password}` (e.g. 'my_user:my_password')
- A kobo API V2 token. See the \href{kobotoolbox document for more information}{https://support.kobotoolbox.org/api.html}}

\item{api}{(character) The type of API to be used or URL at which the API can be accessed. Can be either:
- `humanitarian_response`: "https:://kobo.humanitarianresponse.info" will be used as the base URL (default)
- `kobotoolbox`: "https:://https://kf.kobotoolbox.org" will be used as the base URL
- `custom`: custom URL. Must at least start with "https://"}
}
\value{
`Authorization` header call via \code{\link[httr]{add_headers}}
}
\description{
Retrieves token if not provided in `auth_token`, then add a `Authorization`
header via \code{\link[httr]{add_headers}} to the API call.
}
\details{
All authentication is based on token.
}
