% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboAPI.R
\name{create_export}
\alias{create_export}
\title{Create exports through API}
\usage{
create_export(
  asset_uid,
  user,
  pwd = "",
  api = "https://kobo.humanitarianresponse.info",
  type = "csv",
  lang = "xml",
  fields_from_all_versions = "true",
  hierarchy_in_labels = "false",
  group_sep = "/"
)
}
\arguments{
\item{asset_uid}{UID (ID) of the asset (form) for which an export is created.}

\item{user}{Username of the Kobo account to use.}

\item{api}{The URL at which the API can be accessed.}

\item{type}{Type of exports to create: can be "csv" or "xls". Defaults to "csv"}

\item{lang}{Language to be used for the export. Defaults to "xml"}

\item{fields_from_all_versions}{Include or not all versions of the form. Logical string: "true" or "false". Defaults to "true"}

\item{hierarchy_in_labels}{Group names are displayed or not in labels of the dataframe. Logical string: "true" or "false". Defaults to "false"}

\item{group_sep}{Type of exports to create: can be "csv" or "xls".}

\item{pw}{Password of the Kobo account to use.}
}
\description{
Create exports (endpoint) through API for specified form. Exports allow to download the data via the new KPI.
}
\author{
Punya Prasad Sapkota (https://github.com/ppsapkota/), Elliott Messeiller
}
