% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_utils.R
\name{default_label}
\alias{default_label}
\title{Find default labels in kobo data frames if there is multiple languages.}
\usage{
default_label(questions, language = "english")
}
\arguments{
\item{questions}{A dataframe with questions from a XLSform. Typically the 'survey' sheet.}

\item{language}{The language to use as default if known. Default to english}
}
\value{
A dataframe without groups in headers.
}
\description{
Find default labels in kobo data frames if there is multiple languages.
}
\author{
Elliott Messeiller
}
