% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboAPI.R
\name{download_data}
\alias{download_data}
\title{Download data from the platform}
\usage{
download_data(
  formid,
  user,
  pwd,
  api = "https://kobo.humanitarianresponse.info",
  seperator = "\\\\/"
)
}
\arguments{
\item{formid}{The ID of the form to be accessed (as a character string)}

\item{user}{Optional. A single string indicating the username}

\item{pwd}{Password of the Kobo account to use}

\item{api}{The URL at which the API can be accessed. Default to "kobo.humanitarianresponse.info"}

\item{seperator}{Separator used between select_multiple questions and their choices. Must be a regex expression. Default to forward slash}
}
\value{
A dataframe containing the data, with no group names.
}
\description{
Download data from the platform.
}
\author{
Elliott Messeiller
}
