% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboAPI.R
\name{download_forms_all}
\alias{download_forms_all}
\title{Returns a dataframe with all forms available for the user}
\usage{
download_forms_all(user, pwd, api = "https://kobo.humanitarianresponse.info")
}
\arguments{
\item{user}{Optional. A single string indicating the username}

\item{pwd}{Password of the Kobo account to use}

\item{api}{The URL at which the API can be accessed. Default to "https://kobo.humanitarianresponse.info". "https://kf.kobotoolbox.org" is also supported.}
}
\value{
Dataframe with all forms available for the user.
}
\description{
Download form from the platform
}
\author{
Elliott Messeiller
}
