% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_utils.R
\name{extract_Loops}
\alias{extract_Loops}
\title{Extract repeat loops from the main dataset.}
\usage{
extract_Loops(data, questions, ...)
}
\arguments{
\item{data}{The dataframe to be treated.}

\item{questions}{A dataframe with questions from a XLSform. Typically the 'survey' sheet.}
}
\value{
A dataframe with nested dataframes for all loops.
}
\description{
Extract repeat loops from the main dataset.
}
\author{
Elliott Messeiller
}
