% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-api.R
\name{get_api_url}
\alias{get_api_url}
\title{Get URL of api}
\usage{
get_api_url(api)
}
\arguments{
\item{api}{(character) The type of API to be used or URL at which the API can be accessed. Can be either:
- `humanitarian_response`: "https:://kobo.humanitarianresponse.info" will be used as the base URL (default)
- `kobotoolbox`: "https:://https://kf.kobotoolbox.org" will be used as the base URL
- `custom`: custom URL. Must at least start with "https://"}
}
\value{
character vector with the URL of the api
}
\description{
Get URL of api
}
