% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asset_id.R
\name{get_asset_id_df}
\alias{get_asset_id_df}
\title{Get a data frame with of assets information for the authenticated account.}
\usage{
get_asset_id_df(
  api = c("humanitarian_response", "kobotoolbox"),
  auth_type = c("client", "token"),
  auth_key = NULL
)
}
\arguments{
\item{api}{(character) The type of API to be used or URL at which the API can be accessed. Can be either:
- `humanitarian_response`: "https:://kobo.humanitarianresponse.info" will be used as the base URL (default)
- `kobotoolbox`: "https:://https://kf.kobotoolbox.org" will be used as the base URL
- `custom`: custom URL. Must at least start with "https://"}

\item{auth_type}{(character) Two types of authentication are possible:
- `client`: use the `auth_key` `user:password` to authenticate to the kobo API
- `token`: uses the `token` provided in `auth_key` to authenticate}

\item{auth_key}{(character) Can either be:
- A combination of the kobo and password in format: `{user}:{password}` (e.g. 'my_user:my_password')
- A kobo API V2 token. See the \href{kobotoolbox document for more information}{https://support.kobotoolbox.org/api.html}}
}
\value{
A data frame with asset IDs, names, type, and deployment status.
}
\description{
The function provides a high-level access to APIs based on kobotoolbox. It
parses the base URL for the ´api´ and returns the JSON response in a standard
format.
}
