% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download data of an asset}
\usage{
get_data(
  asset_id,
  api = api,
  auth_type = c("client", "token"),
  auth_key = NULL
)
}
\arguments{
\item{asset_id}{(character) The ID of the form to be accessed (as a character string). Must be a KPI-type ID (not api V1 ID). See download_forms_all for more details.}

\item{api}{(character) The type of API to be used or URL at which the API can be accessed. Can be either:
- `humanitarian_response`: "https:://kobo.humanitarianresponse.info" will be used as the base URL (default)
- `kobotoolbox`: "https:://https://kf.kobotoolbox.org" will be used as the base URL
- `custom`: custom URL. Must at least start with "https://"}

\item{auth_type}{(character) Two types of authentication are possible:
- `client`: use the `auth_key` `user:password` to authenticate to the kobo API
- `token`: uses the `token` provided in `auth_key` to authenticate}

\item{auth_key}{(character) Can either be:
- A combination of the kobo and password in format: `{user}:{password}` (e.g. 'my_user:my_password')
- A kobo API V2 token. See the \href{kobotoolbox document for more information}{https://support.kobotoolbox.org/api.html}}
}
\value{
A data frame with the responses data for ´asset_id´
}
\description{
Download responses from the API for a specific asset
}
