% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKoBo.R
\name{getKoBo}
\alias{getKoBo}
\title{Make an API call.}
\usage{
getKoBo(.url, .username, .password)
}
\arguments{
\item{.url}{can be any URL, for convenience use the object endpoints}

\item{.username}{the username used to authenticate the API call}

\item{.password}{the password used to authenticate the API call}
}
\value{
results of the API call
}
\description{
Make an API call.
}
\examples{
endpoints <- list(assets = "https://kf.kobotoolbox.org/api/v2/assets")
assets <- getKoBo(endpoints$assets, ".username", ".password")
}
