% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboAPI.R
\name{kobo_AddstartCol_SelectMultiple}
\alias{kobo_AddstartCol_SelectMultiple}
\title{Adds a list column with the choices selected at the beggining of select_multiple questions}
\usage{
kobo_AddstartCol_SelectMultiple(data, form, separator = "\\\\/")
}
\arguments{
\item{data}{The dataframe to be treated.}

\item{form}{A list with two objects: The "survey" sheet as a dataframe with all the questions variables, and the "choices" sheet as a dataframe with all the choices variables. See kobo_form()}

\item{separator}{Separator used between select_multiple questions and their choices. Must be a regex expression. Default to forward slash}
}
\value{
Returns data with the additional columns
}
\description{
Adds a list column with the choices selected at the beggining of select_multiple questions
}
\author{
Elliott Messeiller
}
