% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboAPI.R
\name{kobo_form}
\alias{kobo_form}
\title{Download form from the platform}
\usage{
kobo_form(formid, user, api = "https://kobo.humanitarianresponse.info")
}
\arguments{
\item{formid}{The ID of the form to be accessed (as a character string). Must be a KPI-type ID (not api V1 ID). See kobo_all_forms for more details.}

\item{user}{Optional. A single string indicating the username}

\item{api}{The URL at which the API can be accessed. Default to "kobo.humanitarianresponse.info"}
}
\value{
A list with two objects: The "survey" sheet as a dataframe with all the questions variables, and the "choices" sheet as a dataframe with all the choices variables.
}
\description{
Download form from the platform
}
\author{
Elliott Messeiller
}
