% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboAPI.R
\name{new2old_id}
\alias{new2old_id}
\title{Convert form old ID type and return newer ID structure}
\usage{
new2old_id(formid, all_forms)
}
\arguments{
\item{formid}{character string with form id in old format (e.g. 'ay7pkD95i7SLCWy5XkBBZi')}

\item{all_forms}{dataframe containing all the forms linked to the account, as extracted by \link{koboAPI::download_forms_all()}}
}
\value{
a character string with the new format
}
\description{
Convert form old ID type and return newer ID structure
}
