% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_utils.R
\name{noGroupsHeader}
\alias{noGroupsHeader}
\title{Remove groupes from dataframe header}
\usage{
noGroupsHeader(data, questions, separator = "\\\\/")
}
\arguments{
\item{data}{The dataframe to be treated.}

\item{questions}{A dataframe with questions from a XLSform. Typically the 'survey' sheet.}

\item{separator}{Separator used between select_multiple questions and their choices. Must be a regex expression. Default to forward slash}
}
\value{
A dataframe without groups in headers.
}
\description{
Remove groupes from dataframe header
}
\author{
Elliott Messeiller
}
