% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboAPI.R
\name{nullToNA}
\alias{nullToNA}
\title{Null to NA}
\usage{
nullToNA(x)
}
\arguments{
\item{x}{object}
}
\value{
Returns and NA where a NULL was present
}
\description{
Turns NULL blank cells to NAs
}
\author{
Elliott Messeiller
}
