% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koboAPI.R
\name{old2new_id}
\alias{old2new_id}
\title{Convert form old ID type and return newer ID structure}
\usage{
old2new_id(formid, all_forms)
}
\arguments{
\item{formid}{character string with form id in old format (e.g. '390825')}

\item{all_forms}{dataframe containing all the forms linked to the account, as extracted by \link{koboAPI::download_forms_all()}}
}
\value{
a character string with the new format
}
\description{
Convert form old ID type and return newer ID structure
}
\examples{
\dontrun{
old2new_id("390825", all_forms)
}
}
