% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_utils.R
\name{standardize_form}
\alias{standardize_form}
\title{standardize_from}
\usage{
standardize_form(xlsform)
}
\arguments{
\item{xlsform}{xlsform object (list with survey dataframe and choices dataframe)}
}
\value{
Returns xlform object with survey sheet set to standard form
}
\description{
puts xlsform object back in the standard format (same as it was uploaded to server)
}
\author{
Zack Arno
}
