% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nullToNA}
\alias{nullToNA}
\title{Replaces NULL values by NAs}
\usage{
nullToNA(x)
}
\arguments{
\item{x}{Vector to be treated}
}
\value{
Returns the vector with NULL replaced
}
\description{
Replaces NULL values by NAs
}
\author{
Elliott Messeiller
}
